+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added link caching
 - Added named query results caching
 - Added subbody to NamedQuery
+- Added named query first resolver
+- Bug fixes and other small stuff
 
 ## 1.2.5 
 - Support for promises via .fetchSync and .fetchOneSync for client-side queries